pyproject.toml
[tool.poetry]
name = "leaflet"
version = "0.1.0"
description = ""
authors = ["sahil raja <s.raja@imperial.ac.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
dash-leaflet = "^1.0.15"
pandas = "^2.2.3"
plotly = "^5.24.1"
dash-extensions = "^1.0.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
